feat(ui): implement animation in landing page
