+++ package.json
@@ -43,6 +43,7 @@
     "globals": "^16.0.0",
     "husky": "^9.1.7",
     "lint-staged": "^16.0.0",
+    "typescript": "^5.8.3",
     "vite": "^6.3.5"
   }
 }
