+++ src/components/Animation/Animation.css
@@ -1,0 +1,6 @@
+.ydo-animation {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+}
