+++ src/pages/Landing/LandingPage.jsx
@@ -2,23 +2,18 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import "./LandingPage.css";
 import backgroundImage from "../../assets/Landingpage/bg.jpg";
-import arrowImage from "../../assets/Landingpage/arrowImage.svg";
+import YDOAnimation from "../../components/Animation/animation";
 import line from "../../assets/Landingpage/Line.svg";
 
 function Landingpage(){
     const navigate = useNavigate();
 
-    return (
-        <div className="landing_page" style={{ backgroundImage: `url(${backgroundImage})` }}>
+    return (        
+        <div className="landing_page" style={{ backgroundImage: `url(${backgroundImage})` }}>            
             <div className="landing_page_content">
-                <div id="ydo">
-                    <div className="logo_y">Y</div>
-                    <div className="logo_d">D</div>
-                    <div className="image__wrapper">
-                        <img src={arrowImage} alt="arrow" className="arrow" />
-                    </div>
-                    <div className="logo_o">O</div>
-                </div>
+                <div className="animation">
+                    <YDOAnimation />
+                </div>                
                 <p>
                     <div className="line">
                         <img src={line} alt="line" className="Line" />
